Loosen constraints of swipe listener
